[UPDATE][DEBUG] Added new computation problem SOL_R_poly_sys and altered faq
SOL_R_poly_sys deals with finding real solutions to polynomial systems of equations.
Right now, we added Singular and Z3 as possible systems which can tell you something about this.
There was a bug in the faq (wrong folder where the templates are found).
Added that computation problems must - at the current stage - be added to create_tasks_gui.py,
as otherwise an error occurs.

Updated the TODOs.org